spi-tools: do not use git version
Using git version breaks the version when building outside of the
spi-tools git tree.
This can happen when building through an automation tool such as
Buildroot. For this case, a hack is necessary to avoid the git command,
as it performs the build after extracting a tarball and would break
otherwise.
spi-tools was recently added to Buildroot and the necessary hack is
shown in: https://patchwork.ozlabs.org/patch/547120

There is already a VERSION define inside the config.h header generated
by autotools, so my proposal is that this value gets used for the -v
parameter instead of the git return.

Signed-off-by: Erico Nunes <nunes.erico@gmail.com>