Add SPI speed control to spi-pipe.
This commit adds a new -s/--speed option to spi-pipe. This allows
a user to set the SPI speed on a per-command basis without needing
to run spi-config.

This change allows users to experiment more easily with different SPI
speeds, and also enables speed control on platforms where the SPI
speed is reset each time the target spidev is opened or closed.