Remove autoconf AC_FUNC_MALLOC
AC_FUNC_MALLOC does not ensure that a "conforming" malloc() is
available. Instead it just defines malloc to rpl_malloc for which we do
not supply any definition. Remove it, because it does not provide any
benefit this way.

This fixes cross-compilation where runtime-checks cannot be run and
autoconf plays safe and falls back to rpl_malloc, which causes the build
to fail.

Also make sure that malloc is never called with a zero size. glibc
returns NULL in this case, but better avoid doing that and provide a
proper error-message when supplying a blocksize of 0 on command line,
which is the only time malloc is used.